refactor: getFullUserInfo 메서드 이동
