+++ src/test/java/com/dku/council/domain/user/service/UserServiceTest.java
@@ -1,22 +1,16 @@
 package com.dku.council.domain.user.service;
 
-import com.dku.council.domain.comment.repository.CommentRepository;
-import com.dku.council.domain.like.service.LikeService;
-import com.dku.council.domain.post.repository.post.PostRepository;
 import com.dku.council.domain.user.exception.WrongPasswordException;
 import com.dku.council.domain.user.model.UserStatus;
 import com.dku.council.domain.user.model.dto.request.RequestLoginDto;
 import com.dku.council.domain.user.model.dto.request.RequestNickNameChangeDto;
 import com.dku.council.domain.user.model.dto.response.ResponseLoginDto;
-import com.dku.council.domain.user.model.dto.response.ResponseUserInfoDto;
-import com.dku.council.domain.user.model.entity.Major;
 import com.dku.council.domain.user.model.entity.User;
 import com.dku.council.domain.user.repository.UserRepository;
 import com.dku.council.global.auth.jwt.AuthenticationToken;
 import com.dku.council.global.auth.jwt.JwtAuthenticationToken;
 import com.dku.council.global.auth.jwt.JwtProvider;
 import com.dku.council.global.error.exception.UserNotFoundException;
-import com.dku.council.mock.MajorMock;
 import com.dku.council.mock.UserMock;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -28,7 +22,6 @@ import org.springframework.security.crypto.password.PasswordEncoder;
 
 import java.util.Optional;
 
-import static com.dku.council.domain.like.model.LikeTarget.POST;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.mockito.ArgumentMatchers.any;
@@ -47,15 +40,6 @@ class UserServiceTest {
 
     @Mock
     private UserInfoService cacheService;
-
-    @Mock
-    private PostRepository postRepository;
-
-    @Mock
-    private CommentRepository commentRepository;
-
-    @Mock
-    private LikeService likeService;
 
     @Mock
     private JwtProvider jwtProvider;
@@ -133,46 +117,6 @@ class UserServiceTest {
         // then
         assertThat(token.getAccessToken()).isEqualTo("newaccess");
         assertThat(token.getRefreshToken()).isEqualTo("refresh");
-    }
-
-    @Test
-    @DisplayName("내 정보 가져오기")
-    void getUserInfo() {
-        // given
-        Major major = MajorMock.create();
-        User user = UserMock.create(major);
-
-        when(userRepository.findByIdWithMajor(user.getId())).thenReturn(Optional.of(user));
-        when(postRepository.countAllByUserId(user.getId())).thenReturn(1L);
-        when(commentRepository.countAllCommentedByUserId(user.getId())).thenReturn(2L);
-        when(likeService.getCountOfLikedElements(user.getId(), POST)).thenReturn(3L);
-
-        // when
-        ResponseUserInfoDto info = service.getUserInfo(user.getId());
-
-        // then
-        assertThat(info.getStudentId()).isEqualTo(user.getStudentId());
-        assertThat(info.getUsername()).isEqualTo(user.getName());
-        assertThat(info.getNickname()).isEqualTo(user.getNickname());
-        assertThat(info.getYearOfAdmission()).isEqualTo(user.getYearOfAdmission().toString());
-        assertThat(info.getMajor()).isEqualTo(user.getMajor().getName());
-        assertThat(info.getDepartment()).isEqualTo(user.getMajor().getDepartment());
-        assertThat(info.isAdmin()).isEqualTo(user.getUserRole().isAdmin());
-        assertThat(info.getPhoneNumber()).isEqualTo(user.getPhone());
-        assertThat(info.getWritePostCount()).isEqualTo(1);
-        assertThat(info.getCommentedPostCount()).isEqualTo(2);
-        assertThat(info.getLikedPostCount()).isEqualTo(3);
-    }
-
-    @Test
-    @DisplayName("내 정보 가져오기 실패 - 찾을 수 없는 아이디")
-    void failedGetUserInfoByNotFound() {
-        // given
-        when(userRepository.findByIdWithMajor(0L)).thenReturn(Optional.empty());
-
-        // when
-        assertThrows(UserNotFoundException.class, () ->
-                service.getUserInfo(0L));
     }
 
     @Test
